backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20
PyMuPDF>=1.24.0
python-docx>=1.1.0
groq>=0.13.0
python-jobspy>=1.1.0
pydantic>=2.10.0
